api/requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
sqlalchemy==1.4.41
psycopg2-binary==2.9.4
google-cloud-storage==2.5.0
pandas==1.5.0
openpyxl==3.0.10